# Levara/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Levara HNSW + BM25 + Neo4j vector DB HTTP API.\nSee CLAUDE.md for architecture; pkg/mcp for the MCP alternate transport.",
        "title": "Levara API",
        "contact": {},
        "version": "v1"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/cognify": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Transforms text into chunks + embeddings + (optional) graph. Body may provide inline texts[] or reference datasets[] whose raw files are loaded from disk. rag mode skips graph extraction.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cognify"
                ],
                "summary": "Start a cognify pipeline run",
                "parameters": [
                    {
                        "description": "datasets | datasetIds | texts, optional llm_model, collection, session_id",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status + pipeline_run_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "no texts to cognify",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/cognify/{runId}/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cognify"
                ],
                "summary": "Poll the status of a cognify run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run ID returned by POST /cognify",
                        "name": "runId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/runreg.Status"
                        }
                    },
                    "404": {
                        "description": "run not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/cognify/{runId}/stream": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Emits event:progress updates every 500ms while the run is RUNNING, then event:done with the terminal payload. Client should disconnect after the done event.",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "cognify"
                ],
                "summary": "Stream cognify progress via Server-Sent Events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run ID returned by POST /cognify",
                        "name": "runId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE stream",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "run not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "runreg.Status": {
            "type": "object",
            "properties": {
                "chunks_created": {
                    "type": "integer"
                },
                "edges_extracted": {
                    "type": "integer"
                },
                "elapsed_ms": {
                    "type": "integer"
                },
                "entities_extracted": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "pipeline_run_id": {
                    "type": "string"
                },
                "stage": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "description": "RUNNING, COMPLETED, FAILED",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003cjwt\u003e\" from POST /auth/login or an API key from POST /auth/keys.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}